Add gates

+++ capital/app/Http/Controllers/UsersController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\User;
 use App\Models\Role;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class UsersController extends Controller
 {
@@ -61,7 +62,9 @@ class UsersController extends Controller
     public function edit(User $user)
     {
         //
-
+        if (Gate::denies('edit-users')) {
+            return redirect(route('users.index'));
+        }
         $roles = Role::all();
         return view('admin.users.edit')->with([
             'user' => $user,
@@ -93,7 +96,9 @@ class UsersController extends Controller
     public function destroy(User $user)
     {
         //
-
+        if (Gate::denies('delete-users')) {
+            return redirect(route('users.index'));
+        }
         $user->roles()->detach();
         $user->delete();
         return redirect()->route('users.index');
